Fixed an issue in which the nback test would crash because, for some reason, a crucial function was commented out.
Removed translations for print statements, which were going to give an error eventually.

Added a volume test for finding if the participant is comfortable with the current volume.

Now trying to save the settings with illegal inputs will result in the program not saving and giving the user a pop up saying which fields are wrong. Before, the program would save, but eventually crash when attempting to play a test.

Added a check for the function that dealt with string of notes to make sure the developer gets the feedback that the input has to be a string.